Moved SIGNALHUB_GUIDE_URL from env.ts to constants.ts
